# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /orders/{id}:
    get:
      tags:
      - order-rest-controller
      operationId: getOrderByIdReddis
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Order'
...
  /api/orders:
    get:
      tags:
      - order-rest-controller
      summary: Get all orders paged
      operationId: getOrders
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Get orders paged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: No order generated
  /api/orders/{id}/items:
    get:
      tags:
      - order-rest-controller
      summary: Get items of an order
      operationId: getItemsByOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Get items listing of an order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: No item of any order founded
  /api/orders/{ident}:
    get:
      tags:
      - order-rest-controller
      summary: Get an order by id
      operationId: getOrderById
      parameters:
      - name: ident
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Get an order by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: No order founded
  /api/orders/users/{username}:
    get:
      tags:
      - order-rest-controller
      summary: Buy - Generate and order
      operationId: buy
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Generate an order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: No order generated
...
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        gender:
          type: string
        type:
          type: string
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockObject'
        stock:
          $ref: '#/components/schemas/StockObject'
    ItemToBuy:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        size:
          type: string
        count:
          type: integer
          format: int32
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        totalCost:
          type: number
          format: double
        creationDate:
          type: string
        state:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - DELIVERED
          - CANCELLED
        itemToBuys:
          type: array
          items:
            $ref: '#/components/schemas/ItemToBuy'
...
    Size:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        label:
          type: string
    StockObject:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...